add class ArucoDetector, move detectMarkers() and refineDetectedMarkers()
